backend/requirements.txt
streamlit>=1.31
openai>=1.40
python-dotenv>=1.0
pandas>=2.0
openpyxl>=3.1

